Update README.md
To complete all labs in this course, you will need to submit your source for that lab, and demonstrate the lab in operation to one of the course staff (a TA or tutor). You will be graded on the completion of the steps in this lab document. After you have received a signoff, follow the instructions at the end of the document to submit your source code.

Objectives

Develop a realistic real-time application without an operating system

Meet tight timing constraints

Use interrupt prioritization and preemption

Write performance-sensitive code

Access shared data safely

Use real-time debugging techniques